+++ android/src/main/java/com/jiguang/jverify/ResourceUtil.java
@@ -1,0 +1,50 @@
+package com.jiguang.jverify;
+
+import android.content.Context;
+
+/**
+ * Create by wangqingqing
+ * On 2021/4/23 11:13
+ * Copyright(c) 2020 极光
+ * Description
+ */
+public class ResourceUtil {
+
+    public static int getAnimId(Context paramContext, String paramString) {
+        return paramContext.getResources().getIdentifier(paramString, "anim",
+                paramContext.getPackageName());
+    }
+
+    public static int getLayoutId(Context paramContext, String paramString) {
+        return paramContext.getResources().getIdentifier(paramString, "layout",
+                paramContext.getPackageName());
+    }
+
+    public static int getStringId(Context paramContext, String paramString) {
+        return paramContext.getResources().getIdentifier(paramString, "string",
+                paramContext.getPackageName());
+    }
+
+    public static int getDrawableId(Context paramContext, String paramString) {
+        return paramContext.getResources().getIdentifier(paramString,
+                "drawable", paramContext.getPackageName());
+    }
+
+    public static int getStyleId(Context paramContext, String paramString) {
+        return paramContext.getResources().getIdentifier(paramString,
+                "style", paramContext.getPackageName());
+    }
+
+    public static int getId(Context paramContext, String paramString) {
+        return paramContext.getResources().getIdentifier(paramString,"id", paramContext.getPackageName());
+    }
+
+    public static int getColorId(Context paramContext, String paramString) {
+        return paramContext.getResources().getIdentifier(paramString,
+                "color", paramContext.getPackageName());
+    }
+    public static int getArrayId(Context paramContext, String paramString) {
+        return paramContext.getResources().getIdentifier(paramString,
+                "array", paramContext.getPackageName());
+    }
+}